pyproject.toml
[project]
name = "open115"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "httpx[http2]>=0.28.1",
    "pydantic-settings>=2.10.1",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
